- Create ScriptInit for startup - Lots more formatted output for debugging - Allow setting input / output urls - Default read/write is to more convenient directories: -- When script is run with no arguments: -- default in:  WWDC22rawJSON -- default out: docs/wwdc22 -- When using custom paths, they are assumed relative to the script's location
