+++ config/contributors.json
