+++ docs/wwdc22/accessibility-lounge.md
@@ -1,5 +1,6 @@
 # accessibility-lounge QAs
-#### by [FeeTiki](https://twitter.com/FeeTiki)
+### Lounge Contributors
+
 ---
 
 ---
