+++ docs/wwdc22/app-clips-lounge.md
@@ -1,7 +1,9 @@
 # app-clips-lounge QAs
-#### by [shirblc](https://github.com/shirblc)
+### Lounge Contributors
+#### [shirblc](https://github.com/shirblc)
 ---
 
+--- 
 > ####  Due to its size I believe an App Clip should have a limited/basic functionality. How many screens/use-cases/actions should an App Clip provide?
