+++ docs/wwdc22/devtools-and-swift-lounge.md
@@ -1,5 +1,6 @@
 # devtools-and-swift-lounge QAs
-#### by [FeeTiki](https://twitter.com/FeeTiki)
+### Lounge Contributors
+
 ---
 
 ---
