+++ docs/wwdc22/education-lounge.md
@@ -1,7 +1,9 @@
 # education-lounge QAs
-#### by [shirblc](https://github.com/shirblc)
+### Lounge Contributors
+#### [shirblc](https://github.com/shirblc)
 ---
 
+--- 
 > ####  We’re also wondering if there’s any information you’re able to share at this stage about the coming updates to Schoolwork and Classroom apps?
