+++ docs/wwdc22/graphics-and-games-lounge.md
@@ -1,5 +1,6 @@
 # graphics-and-games-lounge QAs
-#### by [FeeTiki](https://twitter.com/FeeTiki)
+### Lounge Contributors
+
 ---
 
 ---
