+++ docs/wwdc22/localization-and-internationalization-lounge.md
@@ -1,7 +1,9 @@
-# Localization and internationalization Lounge QAs
-#### by [emin](https://twitter.com/emin_ui)
+# localization-and-internationalization-lounge QAs
+### Lounge Contributors
+
 ---
 
+--- 
 > ####  Sometimes there is a difference in font size(width/height) between LTR and RTL , how can we make this better?
