+++ docs/wwdc22/photos-camera-lounge.md
@@ -1,5 +1,6 @@
 # photos-camera-lounge QAs
-#### by [FeeTiki](https://twitter.com/FeeTiki)
+### Lounge Contributors
+
 ---
 
 ---
