+++ docs/wwdc22/safari-lounge.md
@@ -1,7 +1,9 @@
 # safari-lounge QAs
-#### by [shirblc](https://github.com/shirblc)
+### Lounge Contributors
+#### [shirblc](https://github.com/shirblc)
 ---
 
+--- 
 > ####  Will the MediaRecorder API be promoted out of the Experimental features in Safari 16?
