+++ docs/wwdc22/swiftui-lounge.md
@@ -1,5 +1,6 @@
 # swiftui-lounge QAs
-#### by [FeeTiki](https://twitter.com/FeeTiki)
+### Lounge Contributors
+
 ---
 
 ---
