+++ docs/wwdc22/uiframeworks-lounge.md
@@ -1,5 +1,6 @@
 # uiframeworks-lounge QAs
-#### by [FeeTiki](https://twitter.com/FeeTiki)
+### Lounge Contributors
+
 ---
 
 ---
